@@ -12,11 +12,12 @@
 //
 //= require jquery
 //= require jquery_ujs
+//= require jquery-ui/tooltip
 //= require dataTables/jquery.dataTables
-//= require dataTables/jquery.dataTables.bootstrap3
-//= require bootstrap.min
+//= require dataTables/bootstrap/3/jquery.dataTables.bootstrap
 //= require d3
 //= require jquery_nested_form
+//= require bootstrap.min
 
 $(document).ready(function() {
   $("[rel~='tooltip']").tooltip();
