@@ -7,9 +7,9 @@ class AdminController < ApplicationController
 
 	def list
 		if params[:approved] == "false"
-			@admins = Admin.sorted.find_all_by_approved(false)
+			@admins = Admin.where(approved: false)
 		else
-			@admins = Admin.sorted
+			@admins = Admin.all
 		end
 	end
 
