@@ -11,11 +11,6 @@ class TemplatesController < ApplicationController
 
 	def show
 		@template = Template.find_by_id(params[:id])
-		if @template.nil?
-			list
-			render('list')
-		end
-
 		@images = @template.images
 	end
 
@@ -35,29 +30,11 @@ class TemplatesController < ApplicationController
 	end
 
 	def edit
-		@template = Template.find_by_id(params[:id])
-		if @template.nil?
-			list
-			render('list')
-		end
+		@template = Template.find(params[:id])
 	end
 
 	def update
 		@template = Template.find(params[:id])
-		attachments = params[:template][:attachments_attributes]
-		unless attachments.nil?
-			attachments.each do |a| 
-				if a[1]["_destroy"].eql? "1"
-					begin
-						@template.attachments.destroy(a[1]["id"])
-						params[:template][:attachments_attributes].delete(a[0])
-					rescue
-						next
-					end
-				end
-			end
-		end
-
 		if @template.update_attributes(params[:template])
 			redirect_to edit_template_path, notice: "Template Updated"
 		else
@@ -74,12 +51,10 @@ class TemplatesController < ApplicationController
 	end
 
 	def destroy
-		# delete folder if_exists?
-		@template = Template.find_by_id(params[:id])
-
-		Template.find(params[:id]).destroy
-		flash[:notice] = "Template Destroyed"
-		redirect_to templates_path
+		@template = Template.find(params[:id])
+		@template.destroy
+		flash[:warning] = "Template Destroyed"
+		redirect_to list_templates_path
 	end
 
 	def copy
@@ -189,9 +164,17 @@ class TemplatesController < ApplicationController
 	def edit_email
 		@attachment = Attachment.find(params[:format])
 		attachment_location = File.join(Rails.root.to_s, "public", "uploads", "attachment", "file", params[:format], "*")
-		@attachment_content = File.read(Dir.glob(attachment_location)[0])
+
+		begin
+			@attachment_content = File.read(Dir.glob(attachment_location)[0])
+		rescue
+			flash[:warning] = "Issue Reading Attachment File"
+			redirect_to :back
+			return
+		end
 		if File.binary?(Dir.glob(attachment_location)[0])
-			redirect_to :back, notice: "Cannot Edit Binary Files"
+			flash[:warning] = "Cannot Edit Binary Files"
+			redirect_to :back
 		end
 	end
 
