@@ -2,7 +2,7 @@ class Victim < ActiveRecord::Base
 	belongs_to :campaign
 	has_many :visits, dependent: :destroy
 
-	validates_format_of :email_address, :with => /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i
+	validates_format_of :email_address, :with => /\A([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})\z/i
 	before_create :default_values
 
 	attr_accessible :email_address, :uid, :campaign_id, :firstname, :lastname
