@@ -2,12 +2,7 @@ require File.expand_path('../boot', __FILE__)
 
 require 'rails/all'
 
-if defined?(Bundler)
-  # If you precompile assets before deploying to production, use this line
-  Bundler.require(*Rails.groups(:assets => %w(development test)))
-  # If you want your assets lazily compiled in production, use this line
-  # Bundler.require(:default, :assets, Rails.env)
-end
+Bundler.require(:default, Rails.env)
 
 module PhishingFramework
   # Default website for reports stuff, must change to callback.
@@ -21,7 +16,7 @@ module PhishingFramework
 
     # Custom directories with classes and modules you want to be autoloadable.
     # config.autoload_paths += %W(#{config.root}/extras)
-
+    
     # Only load the plugins named here, in the order given (default is alphabetical).
     # :all can be used as a placeholder for all plugins not explicitly named.
     # config.plugins = [ :exception_notification, :ssl_requirement, :all ]
@@ -63,7 +58,10 @@ module PhishingFramework
     # Version of your assets, change this if you want to expire all your assets
     config.assets.version = '1.0'
 
-    config.assets.initialize_on_precompile=false
+    config.assets.initialize_on_precompile = false
+
+    # include glyphicons to assets pipeline
+    config.assets.paths << "#{Rails}/vendor/assets/fonts"
 
     # Default url for __Devise__ emails
     config.action_mailer.default_url_options = { :host => 'phishingfrenzy.com' }
